requirements.txt
dnspython
naked
PyInquirer
rich
progressbar
nested_lookup
naked
